Split api folder to manage jobs
