+++ api/Dockerfile.production
@@ -19,6 +19,9 @@ RUN npm run build
 # Run
 FROM node:18-bullseye-slim
 
+# Define build argument with default value 'api'
+ARG SERVER_TYPE=api
+
 WORKDIR /app
 
 ENV NODE_ENV=production
@@ -31,4 +34,4 @@ COPY --from=builder /app/dist ./dist
 
 EXPOSE 8080
 
-CMD ["node", "dist/index.js"]
+CMD node dist/index.js ${SERVER_TYPE}
