+++ api/package.json
@@ -5,13 +5,11 @@
   "main": "src/index.js",
   "scripts": {
     "build": "npx tsc --build --clean && npx tsc -p tsconfig.build.json",
-    "dev": "nodemon",
-    "start-dev": "ts-node src/index.ts",
-    "stop": "pm2 stop api",
-    "start": "npm run build && node dist/index.js",
-    "start-pm2": "pm2 start dist/index.js --name api",
+    "dev:api": "nodemon -- api",
+    "dev:jobs": "nodemon -- jobs",
+    "start:api": "npm run build && node dist/index.js api",
+    "start:jobs": "npm run build && node dist/index.js jobs",
     "prettier": "prettier --write . --plugin=prettier-plugin-organize-imports",
-    "reload": "pm2 reload api",
     "lint": "eslint . --ext .ts,.tsx",
     "lint:fix": "eslint . --ext .ts,.tsx --fix",
     "test": "vitest run",
