+++ api/src/jobs/letudiant/index.ts
@@ -1,15 +1,17 @@
 import { handler } from "./handler";
 import { Queue } from "./queue";
+import { Worker } from "./worker";
+
+export const JOB_NAME = "letudiant-feed";
 
 /**
  * Schedule the feed generation
  * @param cronExpression Expression cron for the schedule (default: every day at 2am)
  */
-async function schedule(cronExpression = "0 2 * * *") {
+export async function schedule(cronExpression = "0 2 * * *") {
   try {
     const queue = Queue.getInstance();
     await queue.scheduleFeedGeneration(cronExpression);
-    console.log(`Letudiant feed generation scheduled with cron: ${cronExpression}`);
   } catch (error) {
     console.error("Failed to schedule letudiant feed generation:", error);
     throw error;
@@ -19,16 +21,17 @@ async function schedule(cronExpression = "0 2 * * *") {
 /**
  * Manually trigger the feed generation
  */
-async function trigger() {
+export async function trigger() {
   try {
     const queue = Queue.getInstance();
     const job = await queue.generateFeed();
-    console.log(`Letudiant feed generation job added with ID: ${job.id}`);
+    console.log(`[Letudiant] Feed generation job added with ID: ${job?.id || 'unknown'}`);
     return job;
   } catch (error) {
-    console.error("Failed to trigger letudiant feed generation:", error);
+    console.error("[Letudiant] Failed to trigger feed generation:", error);
     throw error;
   }
 }
 
-export { handler, schedule, trigger };
+export { Worker, handler };
+
